requirements.txt
python-telegram-bot==20.7
boto3==1.34.8
python-dotenv==1.0.0
loguru==0.7.2
pytz==2023.3.post1
fastapi==0.110.0
uvicorn==0.27.1
starlette==0.36.3
pydantic==2.6.1
playwright>=1.44.0
pyyaml>=6.0
# Фиксированные версии ML-библиотек для совместимости
numpy>=1.20.0,<2.0.0
# Для сервера используем torch==1.13.1
torch==1.13.1
# Для macOS с Apple Silicon можно использовать: torch>=2.0.0
# Фиксируем версию transformers, которая не использует torch.compiler.disable
transformers==4.27.0
psutil>=5.9.0
accelerate==0.30.0 